moving from spec to rpc as requested on pr review
