+++ tests/rpc/rpc_03-chain_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
 
-package spec
+package rpc
 
 import (
 	"os/exec"
@@ -22,11 +22,9 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-
-	rpc "github.com/ChainSafe/gossamer/tests/rpc"
 )
 
-func TestBabeRPC(t *testing.T) {
+func TestChainRPC(t *testing.T) {
 	testsCases := []struct {
 		description string
 		method      string
@@ -34,15 +32,30 @@ func TestBabeRPC(t *testing.T) {
 		skip        bool
 	}{
 		{ //TODO
-			description: "test babe_epochAuthorship",
-			method:      "babe_epochAuthorship",
+			description: "test chain_getHeader",
+			method:      "chain_getHeader",
+			skip:        true,
+		},
+		{ //TODO
+			description: "test chain_getBlock",
+			method:      "chain_getBlock",
+			skip:        true,
+		},
+		{ //TODO
+			description: "test chain_getBlockHash",
+			method:      "chain_getBlockHash",
+			skip:        true,
+		},
+		{ //TODO
+			description: "test chain_getFinalizedHead",
+			method:      "chain_getFinalizedHead",
 			skip:        true,
 		},
 	}
 
-	t.Log("going to Bootstrap Gossamer node")
+	t.Log("going to start gossamer")
 
-	localPidList, err := rpc.StartNodes(t, make([]*exec.Cmd, 1))
+	localPidList, err := StartNodes(t, make([]*exec.Cmd, 1))
 	require.Nil(t, err)
 
 	time.Sleep(time.Second) // give server a second to start
@@ -54,10 +67,10 @@ func TestBabeRPC(t *testing.T) {
 				return
 			}
 
-			respBody, err := rpc.PostRPC(t, test.method, "http://"+rpc.GOSSAMER_NODE_HOST+":"+currentPort, "{}")
+			respBody, err := PostRPC(t, test.method, "http://"+GOSSAMER_NODE_HOST+":"+currentPort, "{}")
 			require.Nil(t, err)
 
-			target := rpc.DecodeRPC(t, respBody, test.method)
+			target := DecodeRPC(t, respBody, test.method)
 
 			require.NotNil(t, target)
 
@@ -66,6 +79,6 @@ func TestBabeRPC(t *testing.T) {
 
 	t.Log("going to TearDown Gossamer node")
 
-	errList := rpc.TearDown(t, localPidList)
+	errList := TearDown(t, localPidList)
 	require.Len(t, errList, 0)
 }
