+++ tests/rpc/rpc_04-offchain_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
 
-package spec
+package rpc
 
 import (
 	"os/exec"
@@ -22,11 +22,9 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-
-	rpc "github.com/ChainSafe/gossamer/tests/rpc"
 )
 
-func TestEngineRPC(t *testing.T) {
+func TestOffchainRPC(t *testing.T) {
 	testsCases := []struct {
 		description string
 		method      string
@@ -34,22 +32,25 @@ func TestEngineRPC(t *testing.T) {
 		skip        bool
 	}{
 		{ //TODO
-			description: "test engine_createBlock",
-			method:      "engine_createBlock",
+			description: "test offchain_localStorageSet",
+			method:      "offchain_localStorageSet",
 			skip:        true,
 		},
 		{ //TODO
-			description: "test engine_finalizeBlock",
-			method:      "engine_finalizeBlock",
+			description: "test offchain_localStorageGet",
+			method:      "offchain_localStorageGet",
+			skip:        true,
+		},
+		{ //TODO
+			description: "test offchain_localStorageGet",
+			method:      "offchain_localStorageGet",
 			skip:        true,
 		},
 	}
 
-	t.Log("going to Bootstrap Gossamer node")
+	t.Log("going to start gossamer")
 
-	localPidList, err := rpc.StartNodes(t, make([]*exec.Cmd, 1))
-
-	//use only first server for tests
+	localPidList, err := StartNodes(t, make([]*exec.Cmd, 1))
 	require.Nil(t, err)
 
 	time.Sleep(time.Second) // give server a second to start
@@ -61,10 +62,10 @@ func TestEngineRPC(t *testing.T) {
 				return
 			}
 
-			respBody, err := rpc.PostRPC(t, test.method, "http://"+rpc.GOSSAMER_NODE_HOST+":"+currentPort, "{}")
+			respBody, err := PostRPC(t, test.method, "http://"+GOSSAMER_NODE_HOST+":"+currentPort, "{}")
 			require.Nil(t, err)
 
-			target := rpc.DecodeRPC(t, respBody, test.method)
+			target := DecodeRPC(t, respBody, test.method)
 
 			require.NotNil(t, target)
 
@@ -73,6 +74,6 @@ func TestEngineRPC(t *testing.T) {
 
 	t.Log("going to TearDown Gossamer node")
 
-	errList := rpc.TearDown(t, localPidList)
+	errList := TearDown(t, localPidList)
 	require.Len(t, errList, 0)
 }
