+++ tests/rpc/rpc_06-engine_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
 
-package spec
+package rpc
 
 import (
 	"os/exec"
@@ -22,11 +22,9 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-
-	rpc "github.com/ChainSafe/gossamer/tests/rpc"
 )
 
-func TestChainRPC(t *testing.T) {
+func TestEngineRPC(t *testing.T) {
 	testsCases := []struct {
 		description string
 		method      string
@@ -34,30 +32,22 @@ func TestChainRPC(t *testing.T) {
 		skip        bool
 	}{
 		{ //TODO
-			description: "test chain_getHeader",
-			method:      "chain_getHeader",
+			description: "test engine_createBlock",
+			method:      "engine_createBlock",
 			skip:        true,
 		},
 		{ //TODO
-			description: "test chain_getBlock",
-			method:      "chain_getBlock",
-			skip:        true,
-		},
-		{ //TODO
-			description: "test chain_getBlockHash",
-			method:      "chain_getBlockHash",
-			skip:        true,
-		},
-		{ //TODO
-			description: "test chain_getFinalizedHead",
-			method:      "chain_getFinalizedHead",
+			description: "test engine_finalizeBlock",
+			method:      "engine_finalizeBlock",
 			skip:        true,
 		},
 	}
 
-	t.Log("going to Bootstrap Gossamer node")
+	t.Log("going to start gossamer")
 
-	localPidList, err := rpc.StartNodes(t, make([]*exec.Cmd, 1))
+	localPidList, err := StartNodes(t, make([]*exec.Cmd, 1))
+
+	//use only first server for tests
 	require.Nil(t, err)
 
 	time.Sleep(time.Second) // give server a second to start
@@ -69,10 +59,10 @@ func TestChainRPC(t *testing.T) {
 				return
 			}
 
-			respBody, err := rpc.PostRPC(t, test.method, "http://"+rpc.GOSSAMER_NODE_HOST+":"+currentPort, "{}")
+			respBody, err := PostRPC(t, test.method, "http://"+GOSSAMER_NODE_HOST+":"+currentPort, "{}")
 			require.Nil(t, err)
 
-			target := rpc.DecodeRPC(t, respBody, test.method)
+			target := DecodeRPC(t, respBody, test.method)
 
 			require.NotNil(t, target)
 
@@ -81,6 +71,6 @@ func TestChainRPC(t *testing.T) {
 
 	t.Log("going to TearDown Gossamer node")
 
-	errList := rpc.TearDown(t, localPidList)
+	errList := TearDown(t, localPidList)
 	require.Len(t, errList, 0)
 }
