+++ tests/rpc/rpc_09-babe_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
 
-package spec
+package rpc
 
 import (
 	"os/exec"
@@ -22,11 +22,9 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-
-	rpc "github.com/ChainSafe/gossamer/tests/rpc"
 )
 
-func TestAuthorRPC(t *testing.T) {
+func TestBabeRPC(t *testing.T) {
 	testsCases := []struct {
 		description string
 		method      string
@@ -34,45 +32,15 @@ func TestAuthorRPC(t *testing.T) {
 		skip        bool
 	}{
 		{ //TODO
-			description: "test author_submitExtrinsic",
-			method:      "author_submitExtrinsic",
-			skip:        true,
-		},
-		{ //TODO
-			description: "test author_pendingExtrinsics",
-			method:      "author_pendingExtrinsics",
-			skip:        true,
-		},
-		{ //TODO
-			description: "test author_removeExtrinsic",
-			method:      "author_removeExtrinsic",
-			skip:        true,
-		},
-		{ //TODO
-			description: "test author_insertKey",
-			method:      "author_insertKey",
-			skip:        true,
-		},
-		{ //TODO
-			description: "test author_rotateKeys",
-			method:      "author_rotateKeys",
-			skip:        true,
-		},
-		{ //TODO
-			description: "test author_hasSessionKeys",
-			method:      "author_hasSessionKeys",
-			skip:        true,
-		},
-		{ //TODO
-			description: "test author_hasKey",
-			method:      "author_hasKey",
+			description: "test babe_epochAuthorship",
+			method:      "babe_epochAuthorship",
 			skip:        true,
 		},
 	}
 
-	t.Log("going to Bootstrap Gossamer node")
+	t.Log("going to start gossamer")
 
-	localPidList, err := rpc.StartNodes(t, make([]*exec.Cmd, 1))
+	localPidList, err := StartNodes(t, make([]*exec.Cmd, 1))
 	require.Nil(t, err)
 
 	time.Sleep(time.Second) // give server a second to start
@@ -84,10 +52,10 @@ func TestAuthorRPC(t *testing.T) {
 				return
 			}
 
-			respBody, err := rpc.PostRPC(t, test.method, "http://"+rpc.GOSSAMER_NODE_HOST+":"+currentPort, "{}")
+			respBody, err := PostRPC(t, test.method, "http://"+GOSSAMER_NODE_HOST+":"+currentPort, "{}")
 			require.Nil(t, err)
 
-			target := rpc.DecodeRPC(t, respBody, test.method)
+			target := DecodeRPC(t, respBody, test.method)
 
 			require.NotNil(t, target)
 
@@ -96,6 +64,6 @@ func TestAuthorRPC(t *testing.T) {
 
 	t.Log("going to TearDown Gossamer node")
 
-	errList := rpc.TearDown(t, localPidList)
+	errList := TearDown(t, localPidList)
 	require.Len(t, errList, 0)
 }
